Moved RefLogPrivate header to Private section
