+++ libGitWrap/Private/RefLogPrivate.hpp
